\name{gfunctions-package}
\alias{gfunctions}
\alias{gfunctions-package}
\docType{package}
\title{\code{glag()} and \code{gsummary()}: Modified versions of \code{lag()} and \code{summary()} }

\description{

\code{\link[gfunctions]{glag}()} and \code{\link[gfunctions]{gsummary}()} are modifications of the \code{\link[stats]{lag}()} and \code{\link[base]{summary}()} functions from the \pkg{stats} package in that they return different information.

}

\details{

\tabular{ll}{
	Version: \tab 1.1\cr
	Date: \tab 2025-07-23\cr
	Licence: \tab GPL-2\cr
}

}

\author{

\tabular{lll}{
	Genaro Sucarrat: \tab \tab \url{https://www.sucarrat.net/}\cr
}

Maintainer: Genaro Sucarrat            

}

\seealso{
\code{\link[stats]{lag}()}, \code{\link[base]{summary}()}
}

\keyword{Statistical Models}
\keyword{Time Series}
\keyword{Econometrics}
\keyword{Financial Econometrics}
