\name{gfunctions-package}
\alias{gfunctions}
\alias{gfunctions-package}
\docType{package}
\title{\code{glag()} and \code{gsummary()}: Modified versions of \code{lag()} and \code{summary()} }

\description{

\code{\link{glag}()} and \code{\link{gsummary}()} are modifications of the \code{\link{lag}()} and \code{\link{summary}()} functions from the \pkg{stats} package in that they return different information.

}

\details{

\tabular{ll}{
	Version: \tab 1.0\cr
	Date: \tab 2024-04-16\cr
	Licence: \tab GPL-2\cr
}

}

\author{

\tabular{lll}{
	Genaro Sucarrat: \tab \tab \url{http://www.sucarrat.net/}\cr
}

Maintainer: Genaro Sucarrat            

}

\seealso{
\code{\link{lag}()}, \code{\link{summary}()}
}

\keyword{Statistical Models}
\keyword{Time Series}
\keyword{Econometrics}
\keyword{Financial Econometrics}
